+++ sdk/python/pulumi_azure_native/servicefabric/__init__.py
@@ -36,7 +36,10 @@ if typing.TYPE_CHECKING:
     import pulumi_azure_native.servicefabric.v20200101preview as v20200101preview
     import pulumi_azure_native.servicefabric.v20200301 as v20200301
     import pulumi_azure_native.servicefabric.v20201201preview as v20201201preview
+    import pulumi_azure_native.servicefabric.v20210101preview as v20210101preview
+    import pulumi_azure_native.servicefabric.v20210501 as v20210501
     import pulumi_azure_native.servicefabric.v20210601 as v20210601
+    import pulumi_azure_native.servicefabric.v20210701preview as v20210701preview
 else:
     v20160901 = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20160901')
     v20170701preview = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20170701preview')
@@ -48,5 +51,8 @@ else:
     v20200101preview = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20200101preview')
     v20200301 = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20200301')
     v20201201preview = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20201201preview')
+    v20210101preview = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20210101preview')
+    v20210501 = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20210501')
     v20210601 = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20210601')
+    v20210701preview = _utilities.lazy_import('pulumi_azure_native.servicefabric.v20210701preview')